hf_space/requirements-hf.txt
# ── CPU-only torch (Python 3.12 compatible) ───────────────────────────
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu

# ── Core ML — exact versions from working Kaggle environment ──────────
transformers==4.57.6
sentence-transformers==5.4.0
accelerate==0.34.2
huggingface_hub==0.36.2

# ── Vector Store ──────────────────────────────────────────────────────
qdrant-client==1.17.1

# ── API ───────────────────────────────────────────────────────────────
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.12.3
python-dotenv==1.0.0
python-multipart==0.0.9

# ── Data ──────────────────────────────────────────────────────────────
pandas==2.1.4
pyarrow==23.0.1
numpy==1.26.4
PyYAML==6.0.1
nltk==3.8.1
tqdm==4.66.3
requests==2.31.0
